Install and export the target
The CMakeLists.txt where a bit update, such that
the project is no build as plain modern cmake project.
Currently, it can be build with in a ROS 1 and ROS 2
workspace via catkin build, or respectively, colcon build.

If you would like to use the lib in a ROS 1 or ROS 2 package,
simply call `find_package(lightfsm)` and use
`target_link_library(<your exe. or lib.> betwo::lightfsm)`.